requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
Pillow==7.1.2
pycparser==2.20
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
